refactor: name and move child component
