+++ src/ts/component/todo/child/todoDetailComponent.ts
@@ -1,11 +1,11 @@
+import { HtmlAccessor } from '../../../htmlUtils/htmlAccessor';
+import { BaseChildComponent } from '../../baseChildComponent';
+import { TemplatePartsHolder } from '../../componentTemplateParts';
 import {
   IDENTIFIER_DETAIL,
-  TODO_DETAIL_NORMAL_HTML,
-  TODO_DETAIL_EDIT_HTML,
-} from '../templates/todoTemplates';
-import { BaseChildComponent } from '../baseChildComponent';
-import { TemplatePartsHolder } from '../componentTemplateParts';
-import { HtmlAccessor } from '../../htmlUtils/htmlAccessor';
+
+  TODO_DETAIL_EDIT_HTML, TODO_DETAIL_NORMAL_HTML
+} from '../../templates/todoTemplates';
 
 export class TodoDetailComponent extends BaseChildComponent {
   constructor(templateParts?: TemplatePartsHolder) {
